Add Twitch Integration
When Galaxe goes live, this will notify the server about the stream and providing a URL and a stream title to players.

This requires a client ID and secret which can be provided from the Twitch Developer panel. You can input those credentials into the config.yml file when launching the server.

Minor changes:

- Added some documentation that was not previously there
- Reordered imports