Modularize some components, add missing documentaiton
Modularized the item meta retrieval due to code complexity concerns, this is now a separate function and will return a pretty formatted TextComponent for the kill message on the server.

Added missing documentation for the event handler as that was missing.

Changed the duration of the Potion effects from infinite time to one hour.