Add documentation, expose hardcoded settings
Added documentation to missing sets of calls for clarity. This will help future development to ensure documentation and knowledge is present.

Exposed settings such as channel name into the configuration file. This is helpful for name changes without making a pull request or commit every time a username is changed.

Updated README to reflect changes with new settings.

Minor changes: 
- Update license header
- Version bump plugin from 1.0-SNAPSHOT to 2.0-SNAPSHOT
- Add JetBrains Annotations for better IDE annotations
- Modified CI job to archive artifacts after build
- Prepping for Qodana code quality checking