Add MIT license, update GitHub Actions, lint project
Added the MIT license for this project as this is a freely open-source project that can be used for reference or any other particular projects.

Updated GitHub Actions to change the cronjob of CodeQL to not have a weird time for execution. Linted the Java build job and added a new Qodana job to assist with code smells.

Code smell fixed with InvisibleItemFrames with a compound assignment.

Updated headers on all Java files.