Add an issue configuration, edit Discord invites
Issue configuration now shows the user to properly disclose bugs with the plugin itself, and not to report issues with the GalaxeSMP itself. This helps cut down on tickets and triaging where there may not be an issue with the current plugin but some other plugin that is not involved with this development.

Updated Discord Invites to use the vanity URL from Discord rather than a randomly generated invite code.