Clean up pronouns command, refactor
Split up each function to make it easier to read and understand.

There are now 2 functions, one for setting pronouns and one for removing pronouns.

Added an enum for pronouns to make it easier to add more pronouns in the future, as well as to make it easier to ensure validity of pronouns.

Added a check for command operations to output useful help messages to the user.

Made changes to use switch statements instead of if statements for better readability.

Added missing plugin.yml entry for the new command.