Add Towny and Graves documentation
Added all related Graves documentation for all the inner workings of how graves spawn.

Added *some* of the extensive Towny documentation as well as a Wiki link to go to for any more information that was not in here.

Removed check for adding white glass pane fillers as its already enabled for every single GUI.

Changed some regular playerheads to be more representative of what they are.