Fix World casting error, startup dependency error
Previously the plugin would throw a ClassCastException when trying to cast a WorldGuard region to a World. This was due to checking if the player had a bypass check and directly casting a normal player to the location, rather than WorldGuard's LocalPlayer. This wasn't documented well from WorldGuard's side, but it is now fixed.

Fixed issue where on startup the plugin would throw an error due to the entirety of the plugin had WorldGuard embedded into it, rather than just the WorldGuard dependency. This is now fixed.

Removed a maven repository due to it conflicting with other dependencies being used.